Separated scoring and ML predicting into own modules, added simple tests for scoring, updated relevant tests and notebook
